+++ ton_client/src/boc/internal.rs
@@ -27,7 +27,7 @@ pub(crate) fn get_boc_hash(boc: &[u8]) -> ClientResult<String> {
     Ok(hex::encode(&id))
 }
 
-pub(crate) fn deserialize_cell_from_base64(
+pub fn deserialize_cell_from_base64(
     b64: &str,
     name: &str,
 ) -> ClientResult<(Vec<u8>, ton_types::Cell)> {
@@ -41,7 +41,7 @@ pub(crate) fn deserialize_cell_from_base64(
     Ok((bytes, cell))
 }
 
-pub(crate) fn deserialize_object_from_cell<S: Deserializable>(
+pub fn deserialize_object_from_cell<S: Deserializable>(
     cell: ton_types::Cell,
     name: &str,
 ) -> ClientResult<S> {
@@ -65,7 +65,7 @@ pub(crate) fn deserialize_object_from_cell<S: Deserializable>(
 }
 
 #[derive(Clone)]
-pub(crate) enum DeserializedBoc {
+pub enum DeserializedBoc {
     Cell(ton_types::Cell),
     Bytes(Vec<u8>),
 }
@@ -80,13 +80,13 @@ impl DeserializedBoc {
 }
 
 #[derive(Clone)]
-pub(crate) struct DeserializedObject<S: Deserializable> {
+pub struct DeserializedObject<S: Deserializable> {
     pub boc: DeserializedBoc,
     pub cell: ton_types::Cell,
     pub object: S,
 }
 
-pub(crate) fn deserialize_object_from_base64<S: Deserializable>(
+pub fn deserialize_object_from_base64<S: Deserializable>(
     b64: &str,
     name: &str,
 ) -> ClientResult<DeserializedObject<S>> {
@@ -100,7 +100,7 @@ pub(crate) fn deserialize_object_from_base64<S: Deserializable>(
     })
 }
 
-pub(crate) fn serialize_object_to_cell<S: Serializable>(
+pub fn serialize_object_to_cell<S: Serializable>(
     object: &S,
     name: &str,
 ) -> ClientResult<ton_types::Cell> {
@@ -109,16 +109,16 @@ pub(crate) fn serialize_object_to_cell<S: Serializable>(
         .map_err(|err| Error::serialization_error(err, name))?)
 }
 
-pub(crate) fn serialize_cell_to_bytes(cell: &ton_types::Cell, name: &str) -> ClientResult<Vec<u8>> {
+pub fn serialize_cell_to_bytes(cell: &ton_types::Cell, name: &str) -> ClientResult<Vec<u8>> {
     ton_types::cells_serialization::serialize_toc(&cell)
         .map_err(|err| Error::serialization_error(err, name))
 }
 
-pub(crate) fn serialize_cell_to_base64(cell: &ton_types::Cell, name: &str) -> ClientResult<String> {
+pub fn serialize_cell_to_base64(cell: &ton_types::Cell, name: &str) -> ClientResult<String> {
     Ok(base64::encode(&serialize_cell_to_bytes(cell, name)?))
 }
 
-pub(crate) fn serialize_object_to_base64<S: Serializable>(
+pub fn serialize_object_to_base64<S: Serializable>(
     object: &S,
     name: &str,
 ) -> ClientResult<String> {
@@ -126,18 +126,14 @@ pub(crate) fn serialize_object_to_base64<S: Serializable>(
     Ok(serialize_cell_to_base64(&cell, name)?)
 }
 
-pub(crate) async fn deserialize_cell_from_boc(
-    context: &ClientContext,
-    boc: &str,
-    name: &str,
+pub async fn deserialize_cell_from_boc(
+    context: &ClientContext, boc: &str, name: &str
 ) -> ClientResult<(DeserializedBoc, ton_types::Cell)> {
     context.bocs.deserialize_cell(boc, name)
 }
 
-pub(crate) async fn deserialize_object_from_boc<S: Deserializable>(
-    context: &ClientContext,
-    boc: &str,
-    name: &str,
+pub async fn deserialize_object_from_boc<S: Deserializable>(
+    context: &ClientContext, boc: &str, name: &str,
 ) -> ClientResult<DeserializedObject<S>> {
     let (boc, cell) = deserialize_cell_from_boc(context, boc, name).await?;
 
@@ -146,7 +142,7 @@ pub(crate) async fn deserialize_object_from_boc<S: Deserializable>(
     Ok(DeserializedObject { boc, cell, object })
 }
 
-pub(crate) fn deserialize_object_from_boc_bin<S: Deserializable>(
+pub fn deserialize_object_from_boc_bin<S: Deserializable>(
     boc: &[u8],
 ) -> ClientResult<(S, UInt256)> {
     let cell =
@@ -157,11 +153,8 @@ pub(crate) fn deserialize_object_from_boc_bin<S: Deserializable>(
     Ok((object, root_hash))
 }
 
-pub(crate) async fn serialize_cell_to_boc(
-    context: &ClientContext,
-    cell: ton_types::Cell,
-    name: &str,
-    boc_cache: Option<BocCacheType>,
+pub async fn serialize_cell_to_boc(
+    context: &ClientContext, cell: ton_types::Cell, name: &str, boc_cache: Option<BocCacheType>,
 ) -> ClientResult<String> {
     if let Some(cache_type) = boc_cache {
         context
@@ -173,11 +166,8 @@ pub(crate) async fn serialize_cell_to_boc(
     }
 }
 
-pub(crate) async fn serialize_object_to_boc<S: Serializable>(
-    context: &ClientContext,
-    object: &S,
-    name: &str,
-    boc_cache: Option<BocCacheType>,
+pub async fn serialize_object_to_boc<S: Serializable>(
+    context: &ClientContext, object: &S,name: &str, boc_cache: Option<BocCacheType>,
 ) -> ClientResult<String> {
     let cell = serialize_object_to_cell(object, name)?;
     serialize_cell_to_boc(context, cell, name, boc_cache).await